Move ner_recognizer to recognizers folder
