+++ lamapi/recognizers/__init__.py
@@ -1,0 +1,6 @@
+"""Recognizers exposed by LamAPI."""
+
+from .literal_recognizer import LiteralRecognizer
+from .ner_recognizer import NERRecognizer
+
+__all__ = ["LiteralRecognizer", "NERRecognizer"]
